Refactor(annotations): Use new annotations from IocUtils
