+++ src/libraries/logger/loggerImpl.ts
@@ -1,19 +1,19 @@
 import TYPES from '../../constants/types';
 import Logger from './logger';
-import { provide, inject } from '../../ioc/ioc';
+import { inject, provideSingleton } from '../../ioc/iocUtils';
 
 /* tslint:disable:no-console */
-@provide(TYPES.Logger)
+@provideSingleton(TYPES.Logger)
 class LoggerImpl implements Logger {
-    private appName: string;
+  private appName: string;
 
-    public constructor(@inject(TYPES.Environment) environment: any) {
-        this.appName = environment.appName;
-    }
+  public constructor(@inject(TYPES.Environment) environment: any) {
+    this.appName = environment.appName;
+  }
 
-    public log(message: string) {
-        console.log(this.appName + ' says: ' + message);
-    }
+  public log(message: string) {
+    console.log(this.appName + ' says: ' + message);
+  }
 }
 
 export default LoggerImpl;
