+++ src/modules/awesomeModule/helloController.ts
@@ -1,14 +1,14 @@
-import { inject, fluentProvide } from '../../ioc/ioc';
+import { fluentProvide, inject } from '../../ioc/iocUtils';
 import Logger from '../../libraries/logger/logger';
 import TYPES from '../../constants/types';
 
 @fluentProvide('Music/HelloController').whenTargetNamed('awesomeModule').done()
 class HelloController {
-    @inject(TYPES.Logger) private logger: Logger;
+  @inject(TYPES.Logger) private logger: Logger;
 
-    public hello(name: string) {
-        this.logger.log('Awesome hello ' + name);
-    }
+  public hello(name: string) {
+    this.logger.log('Awesome hello ' + name);
+  }
 }
 
 export default HelloController;
